+++ tests/test_timepicker.py
@@ -5,8 +5,8 @@ import numpy as np
 import pytest
 from numpy.testing import assert_allclose
 
+from vallenae import timepicker
 from vallenae.io import TraDatabase
-from vallenae.timepicker import aic, energy_ratio, hinkley, modified_energy_ratio
 
 STEEL_PLATE_DIR = Path(__file__).resolve().parent / "../examples/steel_plate"
 SAMPLE_TRADB = STEEL_PLATE_DIR / "sample.tradb"
@@ -21,47 +21,47 @@ def fixture_waveform() -> np.ndarray:
 
 
 def test_hinkley(waveform):
-    result, index = hinkley(waveform, alpha=5)
+    result, index = timepicker.hinkley(waveform, alpha=5)
     assert len(result) == len(waveform)
     assert index == 846
     assert np.argmin(result) == index
-    
+
     # test different alpha value
-    result, index = hinkley(waveform, alpha=10)
+    result, index = timepicker.hinkley(waveform, alpha=10)
     assert index == 837  # A0
 
 
 def test_aic(waveform):
-    def aic_naive(arr: np.ndarray):
-        n = len(arr)
-        result = np.full(n, np.nan, dtype=np.float32)
-        for i in range(n - 1):
-            l_variance = np.var(arr[: i + 1])
-            r_variance = np.var(arr[i + 1:])
-            if l_variance > 0.0 and r_variance > 0.0:
-                result[i] = i * math.log10(l_variance) + (n - i - 1) * math.log10(r_variance)
-            else:
-                result[i] = np.nan
-        return result, np.argmin(result)
-
-    result, index = aic(waveform)
+    result, index = timepicker.aic(waveform)
     assert len(result) == len(waveform)
     assert index == 495  # S0
-
-    result_naive, index_naive = aic(waveform)
-    assert index == index_naive
-    assert_allclose(result, result_naive)
+    assert np.nanargmin(result) == index
 
 
 def test_energy_ratio(waveform):
-    result, index = energy_ratio(waveform, win_len=100)
+    result, index = timepicker.energy_ratio(waveform, win_len=100)
     assert len(result) == len(waveform)
     assert index == 489  # S0
     assert np.argmax(result) == index
 
 
 def test_modified_energy_ratio(waveform):
-    result, index = modified_energy_ratio(waveform, win_len=100)
+    result, index = timepicker.modified_energy_ratio(waveform, win_len=100)
     assert len(result) == len(waveform)
     assert index == 821  # A0
     assert np.argmax(result) == index
+
+
+@pytest.mark.parametrize(
+    ("func_numba, func_numpy"),
+    (
+        (timepicker._hinkley_numba, timepicker._hinkley_numpy),
+        (timepicker._aic_numba, timepicker._aic_numpy),
+        (timepicker._energy_ratio_numba, timepicker._energy_ratio_numpy),
+    ),
+)
+def test_implementations(waveform, func_numba, func_numpy):
+    result_numba, index_numba = func_numba(waveform)
+    result_numpy, index_numpy = func_numpy(waveform)
+    assert index_numba == index_numpy
+    assert_allclose(result_numba, result_numpy, rtol=1e-6, atol=1e-9)